+++ CHANGELOG.md
@@ -17,6 +17,13 @@
     Click to see more.
   </summary>
 
+### :house: Internal
+- `plugin-commons`
+  - [#2608](https://github.com/wix-incubator/rich-content/pull/2608) moved SettingsMobileHeader/UrlInputModal components into ui-components folder
+- `tiptap-editor`
+  - [#2546](https://github.com/wix-incubator/rich-content/pull/2546) package added
+  
+
 </details>
 <hr/>